requirements.txt
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pillow==11.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing_extensions==4.14.1
